+++ tests/specs/externalFormatter/sql.txt
@@ -1,0 +1,6 @@
+-- file.js --
+== should not format SQL tagged template ==
+const query = sql`SELECT   *    FROM users;`;
+
+[expect]
+const query = sql`SELECT   *    FROM users;`;
